mcp/requirements.txt
python-pptx>=1.0.2
mcp
